+++ src/rules/pylint/rules/mod.rs
@@ -9,8 +9,8 @@ pub use merge_isinstance::{merge_isinstance, ConsiderMergingIsinstance};
 pub use nonlocal_without_binding::NonlocalWithoutBinding;
 pub use property_with_parameters::{property_with_parameters, PropertyWithParameters};
 pub use too_many_arguments::{too_many_arguments, TooManyArguments};
-pub use too_many_return_statements::{too_many_return_statements, TooManyReturnStatements};
 pub use too_many_branches::{too_many_branches, TooManyBranches};
+pub use too_many_return_statements::{too_many_return_statements, TooManyReturnStatements};
 pub use too_many_statements::{too_many_statements, TooManyStatements};
 pub use unnecessary_direct_lambda_call::{
     unnecessary_direct_lambda_call, UnnecessaryDirectLambdaCall,
@@ -33,8 +33,8 @@ mod merge_isinstance;
 mod nonlocal_without_binding;
 mod property_with_parameters;
 mod too_many_arguments;
-mod too_many_return_statements;
 mod too_many_branches;
+mod too_many_return_statements;
 mod too_many_statements;
 mod unnecessary_direct_lambda_call;
 mod use_from_import;
